+++ processor/src/test/resources/bad_input/com/example/callback/CallbackNotPublic.java
@@ -1,32 +1,22 @@
 package com.example.callback;
 
-import javax.annotation.Nullable;
-import jsinterop.annotations.JsFunction;
 import react4j.Component;
 import react4j.ReactNode;
 import react4j.annotations.Callback;
 import react4j.annotations.ReactComponent;
 
 @ReactComponent
-abstract class TypeParameterOnCallback
+abstract class CallbackNotPublic
   extends Component
 {
-  @JsFunction
-  public interface CustomHandler
-  {
-    Object render();
-  }
-
-  @Nullable
   @Override
   protected ReactNode render()
   {
     return null;
   }
 
-  @Callback( CustomHandler.class )
-  <T> T handleFoo()
+  @Callback
+  public void handleFoo()
   {
-    return null;
   }
 }